[PSL-713] Add the connectivity check code and if it's not connected, put it in a suspended mode where it won't ban any other nodes
- implemented net manager that runs in a separate thread to check network connectivity
     - different check wait time for active/inactive modes
     - checks for active network interface (Windows,Linux specific versions) and for Internet connectivity
- changed logic that checks for nodes inactivity and disconnects (ban) nodes:
    - skip inactive nodes check if network is unreachable
   - use grace period for timed out nodes when network comes up, force nodes to send ping message again
- changed some Node fields related to ping functionality to atomic types (fields that can be accessed simultaneously from different threads):
- fixed some compilation warnings for missing override in virtual methods

+++ src/rpc/net.cpp
@@ -1,6 +1,7 @@
 // Copyright (c) 2009-2014 The Bitcoin Core developers
+// Copyright (c) 2018-2023 The Pastel Core developers
 // Distributed under the MIT software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
+// file COPYING or https://www.opensource.org/licenses/mit-license.php.
 
 #include <univalue.h>
 
@@ -156,8 +157,9 @@ Examples:
         // their ver message.
         obj.pushKV("subver", stats.cleanSubVer);
         obj.pushKV("inbound", stats.fInbound);
-        obj.pushKV("startingheight", stats.nStartingHeight);
-        if (fStateStats) {
+        obj.pushKV("startingheight", static_cast<uint64_t>(stats.nStartingHeight));
+        if (fStateStats)
+        {
             obj.pushKV("banscore", statestats.nMisbehavior);
             obj.pushKV("synced_headers", statestats.nSyncHeight);
             obj.pushKV("synced_blocks", statestats.nCommonHeight);
